graphql code restructure and organization (#3895)
* basic user friend resolver

* restructure resolvers to make them more managable
