+++ backend/native/backpack-api/src/routes/graphql/resolvers/index.ts
@@ -1,0 +1,2 @@
+export * from "./user";
+export * from "./wallet";
